Fixed html-minifier overwrites source files and outputs empty files.
html-minifier 3.2.3 doesn't make a copy of the source file even if the output file given with -o flag is the same as the source file, and overwrites the source file. As the result of this behavior, html-minifier reads a blank file and outputs nothing.

This commit updates the command running html-minifier to assign a different inode number to the output file, which means the source file and the output file become pointing different inodes and html-minifier doesn't overwrite the source file.